fix-bug
corrigidos os erros

+++ src/main/java/com/OptimizationBlood/blood/repository/SangueRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SangueRepository extends JpaRepository<Sangue,Integer> {
 
-    Sangue findByCodigo(Long codigo);
+    Sangue findByCodigo(int codigo);
 }
